requirements.txt
pillow>=9.0.0
numpy>=1.20.0
rich>=10.0.0
cryptography>=41.0.0